Thus far these are tested and working
they also contain some hints and tips and notes to familiarize yourself
with the code should you want to do some bugfixing.

I ask you do not change the code completely. If you want to start again,
use a different file. Fixes and such are alright as long as they are
documented .-.